refactor: unify terminal runners from cli or api
