+++ packages/skott/test.mjs
@@ -4,19 +4,20 @@ import { Terminal, Web } from "./dist/src/rendering/api.js"
 
 
 
-// Terminal.renderTerminalApplication(defaultConfig, { displayMode: "graph", watch: true })
+Terminal.renderTerminalApplication(defaultConfig, { displayMode: "webapp", watch: true })
 
-Web.renderWebApplication(defaultConfig, {
-    open: true,
-    port: 1111,
-    visualization: {
-        granularity: "module"
-    },
-    onListen: () => {
-        console.log("Listening")
-    },
-    onOpen: () => {
-        console.log("Open")
-    }
-})
+// Web.renderWebApplication(defaultConfig, {
+//     open: true,
+//     watch: true,
+//     port: 1111,
+//     visualization: {
+//         granularity: "module"
+//     },
+//     onListen: () => {
+//         console.log("Listening")
+//     },
+//     onOpen: () => {
+//         console.log("Open")
+//     }
+// })
